requirements.txt
Django>=2.2.1,<2.3.0
djangorestframework>=3.9.3,<3.10.0
flake8>=3.7.7,<3.8.0
psycopg2>=2.8.2,<2.9.0
Faker>=1.0.5,<1.1.0
django-extensions>=2.1.6,<2.2.0
